refactor: :recycle: refactor ActorDataManager
